[ref-perf] refactor user-basic-info hook to be used only by the presenter.
