+++ src/components/modal/types.ts
@@ -1,4 +1,4 @@
-import { CurrentUserData, DeleteEntryFunction, GraphqlResponseWrapper } from 'bigbluebutton-html-plugin-sdk';
+import { CurrentUserData, DeleteEntryFunction, GraphqlResponseWrapper, PluginApi } from 'bigbluebutton-html-plugin-sdk';
 import { IntlShape } from 'react-intl';
 import { DataChannelEntryResponseType, PushEntryFunction } from 'bigbluebutton-html-plugin-sdk/dist/cjs/data-channel/types';
 import { PickedUser, PickedUserWithEntryId, PickedUserSeenEntryDataChannel } from '../pick-random-user/types';
@@ -10,9 +10,8 @@ export interface PickUserModalProps {
   intl: IntlShape
   showModal: boolean;
   handleCloseModal: () => void;
-  users?: PickedUser[];
-  currentPickedUser: PickedUserWithEntryId;
-  handlePickRandomUser: () => void;
+  pluginApi: PluginApi;
+  currentPickedUser: PickedUserWithEntryId | null;
   currentUser: CurrentUserData;
   dataChannelPickedUsers?: DataChannelEntryResponseType<PickedUser>[];
   deletionFunction: DeleteEntryFunction;
