+++ src/components/pick-random-user/hooks.ts
@@ -1,10 +1,7 @@
 import {
   CurrentUserData,
   DataChannelEntryResponseType,
-  DataChannelTypes,
   GraphqlResponseWrapper,
-  PluginApi,
-  PushEntryFunction,
 } from 'bigbluebutton-html-plugin-sdk';
 import { PluginSettingsData } from 'bigbluebutton-html-plugin-sdk/dist/cjs/data-consumption/domain/settings/plugin-settings/types';
 import { useEffect, useState } from 'react';
@@ -17,7 +14,6 @@ import { hasCurrentUserSeenPickedUser, isNumber } from '../../commons/utils';
 import { PickRandomUserSettings } from '../../commons/types';
 import { WindowClientSettings } from '../modal/types';
 import {
-  FilterOptionsType,
   PickedUser,
   PickedUserSeenEntryDataChannel,
   PickedUserWithEntryId,
@@ -125,95 +121,14 @@ export const useGetAllSettings = (
   };
 };
 
-// Filter Options hooks and utilities
-
-const useUpdateFilterOptionsOnDataChannel = (
-  pushFilterOptionsToDataChannel: PushEntryFunction<FilterOptionsType>,
-  filterOptions: FilterOptionsType,
-  isPresenter: boolean,
-  dataChannelLoading: boolean,
-) => {
-  useEffect(() => {
-    if (isPresenter && !dataChannelLoading) {
-      pushFilterOptionsToDataChannel(filterOptions);
-    }
-  }, [isPresenter, filterOptions, dataChannelLoading]);
-};
-
-const hasFilterOptionsChanged = (
-  currentFilterOptions: FilterOptionsType,
-  filterOptionsFromDataChannel?: FilterOptionsType,
-) => filterOptionsFromDataChannel?.includePickedUsers !== currentFilterOptions.includePickedUsers
-  || filterOptionsFromDataChannel?.includeModerators !== currentFilterOptions.includeModerators
-  || filterOptionsFromDataChannel?.includePresenter !== currentFilterOptions.includePresenter;
-
-const useObserveFilterOptionsFromDataChannel = (
-  currentFilterOptions: FilterOptionsType,
-  filterOptionsFromDataChannel: FilterOptionsType | undefined,
-  setFilterOptions: React.Dispatch<React.SetStateAction<FilterOptionsType>>,
-) => {
-  useEffect(() => {
-    if (
-      filterOptionsFromDataChannel
-      && hasFilterOptionsChanged(currentFilterOptions, filterOptionsFromDataChannel)) {
-      setFilterOptions({
-        includePickedUsers: filterOptionsFromDataChannel.includePickedUsers,
-        includeModerators: filterOptionsFromDataChannel.includeModerators,
-        includePresenter: filterOptionsFromDataChannel.includePresenter,
-      });
-    }
-  }, [filterOptionsFromDataChannel]);
-};
-
-const getLatestFilterOptionsFromDataChannel = (
-  filterOptionsFromDataChannelResponse: GraphqlResponseWrapper<
-    DataChannelEntryResponseType<FilterOptionsType>[]
-  >,
-) => {
-  const persistedFilterOptionsList = filterOptionsFromDataChannelResponse.data;
-  const currentFilterOptionsFromDataChannel = persistedFilterOptionsList
-    ? persistedFilterOptionsList[0]?.payloadJson : null;
-  return currentFilterOptionsFromDataChannel;
-};
-
-export const useGetFilterOptions = (
-  pluginApi: PluginApi,
-  currentUserPresenter: boolean,
-): [FilterOptionsType, React.Dispatch<React.SetStateAction<FilterOptionsType>>] => {
-  const [filterOptions, setFilterOptions] = useState<FilterOptionsType>({
-    includeModerators: false,
-    includePresenter: false,
-    includePickedUsers: false,
-  });
-  const {
-    data: filterOptionsFromDataChannel,
-    pushEntry: pushFilterOptionsToDataChannel,
-  } = pluginApi.useDataChannel<FilterOptionsType>('filterOptions', DataChannelTypes.LATEST_ITEM);
-  const latestFilterOptionFromDataChannel = getLatestFilterOptionsFromDataChannel(
-    filterOptionsFromDataChannel,
-  );
-  useObserveFilterOptionsFromDataChannel(
-    filterOptions,
-    latestFilterOptionFromDataChannel,
-    setFilterOptions,
-  );
-  useUpdateFilterOptionsOnDataChannel(
-    pushFilterOptionsToDataChannel,
-    filterOptions,
-    currentUserPresenter,
-    filterOptionsFromDataChannel.loading,
-  );
-  return [filterOptions, setFilterOptions];
-};
-
 // ---
 
 export const useGetCurrentPickedUser = (
-  pickedUserFromDataChannel: DataChannelEntryResponseType<PickedUser>[],
-): PickedUserWithEntryId | undefined => {
+  pickedUserFromDataChannel: DataChannelEntryResponseType<PickedUser>[] | undefined,
+): PickedUserWithEntryId | null => {
   const [
     pickedUserWithEntryId,
-    setPickedUserWithEntryId] = useState<PickedUserWithEntryId | undefined>();
+    setPickedUserWithEntryId] = useState<PickedUserWithEntryId | null>(null);
 
   useEffect(() => {
     if (pickedUserFromDataChannel
@@ -239,7 +154,7 @@ export const useControlModalState = (
     DataChannelEntryResponseType<PickedUserSeenEntryDataChannel>[]>,
   currentUser: CurrentUserData,
   pickedUserTimeWindow: number,
-  currentPickedUser: PickedUserWithEntryId,
+  currentPickedUser: PickedUserWithEntryId | null,
   setShowModal: React.Dispatch<React.SetStateAction<boolean>>,
 ) => {
   const hasValidPickInTimeWindow = (pickedUser?: DataChannelEntryResponseType<PickedUser>) => {
